Replace custom DataTableSkeleton component with Carbon version
The Carbon v11 DataTableSkeleton is now much more configurable than
the old Carbon v10 version. There is no need for our own custom
component and styling anymore.
