--- packages/components/src/components/DataTableSkeleton/_DataTableSkeleton.scss
@@ -1,54 +1,0 @@
-/*
-Copyright 2020-2024 The Tekton Authors
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-@use '@carbon/react/scss/config' as *;
-@use '@carbon/react/scss/utilities/skeleton' as *;
-@use '@carbon/react/scss/theme' as *;
-// TODO: carbon11 - check Carbon 11 DataTable skeleton state to see if it's suitable now
-
-.tkn--data-table-skeleton.#{$prefix}--data-table.#{$prefix}--skeleton {
-  th {
-    vertical-align: middle;
-
-    &:nth-child(3n + 1) {
-      inline-size: 10%;
-    }
-
-    &:nth-child(3n + 2) {
-      inline-size: 30%;
-    }
-
-    &:nth-child(3n + 3) {
-      inline-size: 15%;
-    }
-  }
-
-  th span,
-  td span {
-    @include skeleton;
-    inline-size: 75%;
-    block-size: 1rem;
-    display: block;
-  }
-
-  tr:hover {
-    td {
-      border-color: $border-subtle;
-      background: transparent;
-      &:first-of-type,
-      &:last-of-type {
-        border-color: $border-subtle;
-      }
-    }
-  }
-}
