+++ src/scss/App.scss
@@ -32,7 +32,6 @@ limitations under the License.
 
 @use '@tektoncd/dashboard-components/src/scss/common';
 @use '@tektoncd/dashboard-components/src/scss/Run';
-@use '@tektoncd/dashboard-components/src/components/DataTableSkeleton/DataTableSkeleton';
 @use '@tektoncd/dashboard-components/src/components/DeleteModal/DeleteModal';
 @use '@tektoncd/dashboard-components/src/components/DetailsHeader/DetailsHeader';
 @use '@tektoncd/dashboard-components/src/components/DotSpinner/DotSpinner';
