+++ src/scss/_carbon.scss
@@ -56,7 +56,6 @@ limitations under the License.
 @use '@carbon/react/scss/components/pagination';
 @use '@carbon/react/scss/components/radio-button';
 @use '@carbon/react/scss/components/search';
-@use '@carbon/react/scss/components/skeleton-styles';
 @use '@carbon/react/scss/components/stack';
 @use '@carbon/react/scss/components/tabs';
 @use '@carbon/react/scss/components/tag';
